requirements.txt
statistics
pytest
pyyaml